requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
python-telegram-bot>=20.0
ccxt>=4.0.0
matplotlib>=3.7.0
python-dotenv>=0.21.0
websockets>=11.0.3
python-dateutil>=2.8.2
ujson>=5.8.0
cryptography>=41.0.0
colorama>=0.4.6
plotly>=5.17.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2
pytest>=7.4.0
pandas-ta
